Rename ComputeNodeStateChangedWatcher
